pyproject.toml
[project]
name = "ms-stt"
version = "0.1.0"
description = "A microservice that converts speech to text"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.76.0",
    "numpy>=2.2.6",
    "pydantic>=2.12.5",
    "sounddevice>=0.5.3",
    "structlog>=25.5.0",
    "vosk>=0.3.45",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
    "rich>=14.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ms_stt"]
